requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.3
django-admin-interface==0.30.1
django-colorfield==0.14.0
django-environ==0.12.0
django-grappelli==4.0.2
django-jazzmin==3.0.1
django-jet==1.0.8
django-perms==2.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mysqlclient==2.2.7
pillow==11.2.1
PyJWT==2.9.0
python-slugify==8.0.4
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
uritemplate==4.2.0
